+++ ureport/backend/tests/test_rapidpro.py
@@ -231,7 +231,8 @@ class ContactSyncerTest(UreportTest):
                                             fields={'registration_date': None, 'state': None,
                                                     'lga': None, 'occupation': None, 'born': None,
                                                     'gender': None},
-                                            language='eng')
+                                            language='eng',
+                                            created_on=json_date_to_datetime('2013-01-02T03:04:05.000'))
 
         self.assertEqual(self.syncer.local_kwargs(self.nigeria, temba_contact),
                          {'backend': self.rapidpro_backend,
@@ -240,7 +241,7 @@ class ContactSyncerTest(UreportTest):
                           'gender': '',
                           'born': 0,
                           'occupation': '',
-                          'registered_on': None,
+                          'registered_on': json_date_to_datetime('2013-01-02T03:04:05.000'),
                           'state': '',
                           'district': '',
                           'ward': ''})
@@ -448,19 +449,22 @@ class RapidProBackendTest(UreportTest):
                     TembaContact.create(
                         uuid="C-001", name="Bob McFlow", language="eng", urns=["twitter:bobflow"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "34"}, stopped=False, blocked=False
+                        fields={'age': "34"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                     TembaContact.create(
                         uuid="C-002", name="Jim McMsg", language="fre", urns=["tel:+250783835665"],
                         groups=[ObjectRef.create(uuid="G-002", name="Spammers")],
-                        fields={'age': "67"}, stopped=False, blocked=False
+                        fields={'age': "67"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ],
                 [
                     TembaContact.create(
                         uuid="C-003", name="Ann McPoll", language="eng", urns=["tel:+250783835664"],
                         groups=[],
-                        fields={'age': "35"}, stopped=True, blocked=False
+                        fields={'age': "35"}, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             ),
@@ -469,7 +473,8 @@ class RapidProBackendTest(UreportTest):
                 [
                     TembaContact.create(
                         uuid="C-004", name=None, language=None, urns=[], groups=[],
-                        fields=None, stopped=True, blocked=False
+                        fields=None, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             )
@@ -489,19 +494,22 @@ class RapidProBackendTest(UreportTest):
                     TembaContact.create(
                         uuid="C-001", name="Bob McFlow", language="eng", urns=["twitter:bobflow"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "34"}, stopped=False, blocked=False
+                        fields={'age': "34"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                     TembaContact.create(
                         uuid="C-002", name="Jim McMsg", language="fre", urns=["tel:+250783835665"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "67"}, stopped=False, blocked=False
+                        fields={'age': "67"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ],
                 [
                     TembaContact.create(
                         uuid="C-003", name="Ann McPoll", language="eng", urns=["tel:+250783835664"],
                         groups=[],
-                        fields={'age': "35"}, stopped=True, blocked=False
+                        fields={'age': "35"}, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             ),
@@ -510,7 +518,8 @@ class RapidProBackendTest(UreportTest):
                 [
                     TembaContact.create(
                         uuid="C-004", name=None, language=None, urns=[], groups=[],
-                        fields=None, stopped=True, blocked=False
+                        fields=None, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             )
@@ -531,19 +540,22 @@ class RapidProBackendTest(UreportTest):
                     TembaContact.create(
                         uuid="C-001", name="Bob McFlow", language="eng", urns=["twitter:bobflow"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "34"}, stopped=False, blocked=False
+                        fields={'age': "34"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                     TembaContact.create(
                         uuid="C-002", name="Jim McMsg", language="fre", urns=["tel:+250783835665"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "67"}, stopped=False, blocked=False
+                        fields={'age': "67"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ],
                 [
                     TembaContact.create(
                         uuid="C-003", name="Ann McPoll", language="eng", urns=["tel:+250783835664"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "35"}, stopped=True, blocked=False
+                        fields={'age': "35"}, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             ),
@@ -552,7 +564,8 @@ class RapidProBackendTest(UreportTest):
                 [
                     TembaContact.create(
                         uuid="C-004", name=None, language=None, urns=[], groups=[],
-                        fields=None, stopped=True, blocked=False
+                        fields=None, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             )
@@ -577,11 +590,12 @@ class RapidProBackendTest(UreportTest):
                                         fields={'registration_date': '2014-01-02T03:04:05.000000Z',
                                                 'state': 'Nigeria > Lagos', 'lga': 'Nigeria > Lagos > Oyo',
                                                 'occupation': 'Student', 'born': '1990', 'gender': 'Male'},
-                                        language='eng'),
+                                        language='eng', created_on='2013-01-02T03:04:05.000000Z'),
                     TembaContact.create(
                         uuid="C-002", name="Jim McMsg", language="fre", urns=["tel:+250783835665"],
                         groups=[ObjectRef.create(uuid="G-001", name="ureporters")],
-                        fields={'age': "67", "born": "1992"}, stopped=False, blocked=False
+                        fields={'age': "67", "born": "1992"}, stopped=False, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             ),
@@ -590,7 +604,8 @@ class RapidProBackendTest(UreportTest):
                 [
                     TembaContact.create(
                         uuid="C-004", name=None, language=None, urns=[], groups=[],
-                        fields=None, stopped=True, blocked=False
+                        fields=None, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             )
@@ -618,7 +633,8 @@ class RapidProBackendTest(UreportTest):
                 [
                     TembaContact.create(
                         uuid="C-002", name=None, language=None, urns=[], groups=[],
-                        fields=None, stopped=True, blocked=False
+                        fields=None, stopped=True, blocked=False,
+                        created_on=json_date_to_datetime('2013-01-02T03:04:05.000'),
                     ),
                 ]
             )
@@ -936,6 +952,41 @@ class RapidProBackendTest(UreportTest):
 
         self.assertEqual((num_val_created, num_val_updated, num_val_ignored,
                           num_path_created, num_path_updated, num_path_ignored), (0, 1, 0, 0, 0, 3))
+
+        long_text = "Long Text moree " * 100000
+        temba_run_long_text = TembaRun.create(id=1234, flow=ObjectRef.create(uuid='flow-uuid', name="Flow 1"),
+                                              contact=ObjectRef.create(uuid='C-007', name='Lebron James'), responded=True,
+                                              values={"party": TembaRun.Value.create(value=long_text,
+                                                                                     category="Party",
+                                                                                     node='ruleset-uuid',
+                                                                                     time=now + timedelta(minutes=1))},
+                                              path=[TembaRun.Step.create(node='ruleset-uuid', time=now)],
+                                              created_on=now, modified_on=now, exited_on=now,
+                                              exit_type='completed')
+
+        mock_get_runs.side_effect = [MockClientQuery([temba_run_long_text])]
+        (num_val_created, num_val_updated, num_val_ignored,
+         num_path_created, num_path_updated, num_path_ignored) = self.backend.pull_results(poll, None, None)
+
+        self.assertEqual((num_val_created, num_val_updated, num_val_ignored,
+                          num_path_created, num_path_updated, num_path_ignored), (1, 0, 0, 0, 0, 1))
+
+        temba_run_none_value = TembaRun.create(id=1234, flow=ObjectRef.create(uuid='flow-uuid', name="Flow 1"),
+                                               contact=ObjectRef.create(uuid='C-008', name='James Harden'), responded=True,
+                                               values={"party": TembaRun.Value.create(value=None,
+                                                                                      category="Party",
+                                                                                      node='ruleset-uuid',
+                                                                                      time=now + timedelta(minutes=1))},
+                                               path=[TembaRun.Step.create(node='ruleset-uuid', time=now)],
+                                               created_on=now, modified_on=now, exited_on=now,
+                                               exit_type='completed')
+
+        mock_get_runs.side_effect = [MockClientQuery([temba_run_none_value])]
+        (num_val_created, num_val_updated, num_val_ignored,
+         num_path_created, num_path_updated, num_path_ignored) = self.backend.pull_results(poll, None, None)
+
+        self.assertEqual((num_val_created, num_val_updated, num_val_ignored,
+                          num_path_created, num_path_updated, num_path_ignored), (1, 0, 0, 0, 0, 1))
 
     @patch('dash.orgs.models.TembaClient.get_runs')
     @patch('django.utils.timezone.now')